Move lazy evaluator include operation classes to separate file
